# perf: Tweak slot calculation algorithm