The previous slot calculation algorithm we use is optimized for large sequential programs. While this works for simpler cases, it often results in address conflicts in terms of programs with frequent branches. This PR changes the code to a new algorithm that balances between small basic blocks and large code segments. For our most common secp256k1 verification, this reduces ~15% of running time:

```
Benchmarking interpret secp256k1_bench via assembly: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 12.7s or reduce sample count to 40
interpret secp256k1_bench via assembly
                        time:   [2.5024 ms 2.5031 ms 2.5038 ms]
                        change: [-16.289% -16.251% -16.212%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  5 (5.00%) high mild
  2 (2.00%) high severe
```

For more complicated programs, such as [bn256g2](https://github.com/nervosnetwork/ckb-miscellaneous-scripts/blob/zk/c/bn256g2.hpp) where branches are more frequent, we noticed a 4x speedup.